+++ js/main.js
@@ -1,3 +1,2 @@
-import {generatedPhotos} from './generate-photos.js';
-
-//console.log(generatedPhotos);
+import {renderMiniature} from './miniatures.js';
+renderMiniature();
