+++ js/util.js
@@ -12,5 +12,6 @@ const getRandomInteger = function (min, max) {
 const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
 //Проверка длины строки:
 const checkStringLength = (string, Length) => string.length <= Length;
+//console.log(checkStringLength('What happened to Gul\'dan after Illidan took his cane? He fell.', 69));
 
 export {getRandomInteger, getRandomArrayElement, checkStringLength};
